phase-C-2: refactor(bin): replace hand-rolled invalid-cmd check with yargs strictCommands
The bin had ~20 lines that re-implemented what yargs.strictCommands()
already does: detect an unknown command, show help, exit non-zero. The
hand-rolled version also short-circuited --help/--version recording
into metrics, which was actually a bug — it recorded "--help" as a
command name in the daily usage stats.

Replace it with .strictCommands().recommendCommands() and a single
recordMetrics call after parse(). On unknown commands, yargs now
prints help + "Did you mean cget?" — strictly better UX than the
old "Invalid command provided" line.

Also moved the createMysqlAction helper from src/actions/mysql/ to
src/utils/mysql/. The previous location got picked up by action
discovery, since the discoverer treats any .js file in actions/ dirs
as an action factory. Helpers belong with utils, not actions.

Drops three lodash imports from bin (toString, find, isEmpty); union
stays for one more commit and goes in phase-C-3.

Bin spec updated to match yargs's new "Did you mean ..." wording.
